+++ crates/icydb-core/src/db/sql/lowering/predicate/tests.rs
@@ -114,7 +114,7 @@ fn derive_where_predicate_subset_recovers_folded_constant_compare_shapes() {
 fn compile_where_bool_expr_requires_normalized_shape() {
     let expr = Expr::Binary {
         op: crate::db::query::plan::expr::BinaryOp::Eq,
-        left: Box::new(Expr::Literal(Value::Int(5))),
+        left: Box::new(Expr::Literal(Value::Int64(5))),
         right: Box::new(Expr::Field(FieldId::new("age"))),
     };
